+++ sdmx-dl-provider-dialects/src/main/java/module-info.java
@@ -20,6 +20,7 @@ module sdmxdl.provider.dialects {
             DotStatDialectDriver,
             EstatDialectDriver,
             ImfDialectDriver,
+            IneDialectDriver,
             InseeDialectDriver,
             StatCanDialectDriver,
             UisDialectDriver;